# HTNest/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "NZ Walks API",
    "version": "V1"
  },
  "paths": {
...
    "/api/Product": {
      "get": {
        "tags": [
          "Product"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "nameProduct",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "ProductCode",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Image",
            "required": true,
            "type": "file"
          },
          {
            "in": "formData",
            "name": "ProductName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Description",
            "required": true,
            "type": "string",
            "pattern": "^[a-zA-Z0-9\\s]*$"
          },
          {
            "in": "formData",
            "name": "StockQuantity",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Discount",
            "type": "number",
            "format": "double",
            "maximum": 100,
            "minimum": 1
          },
          {
            "in": "formData",
            "name": "Ingredient",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Warning",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "CategoryId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "StyleId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Price",
            "required": true,
            "type": "number",
            "format": "double",
            "minimum": 10,
            "pattern": "^\\d+(\\.\\d{1,2})?$"
          },
          {
            "in": "formData",
            "name": "Volume",
            "type": "number",
            "format": "double"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        }
      }
    },
    "/api/Product/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateProductModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        }
      }
    },
    "/api/Style": {
      "get": {
        "tags": [
          "Style"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "styleName",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Style"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Style"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateStyleModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Style"
            }
          }
        }
      }
    },
    "/api/Style/{id}": {
      "get": {
        "tags": [
          "Style"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Style"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Style"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateStyleModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Style"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Style"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Style"
            }
          }
        }
      }
    },
    "/api/User": {
      "get": {
        "tags": [
          "User"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateUserModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/User/{userName}": {
      "get": {
        "tags": [
          "User"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userName",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateUserModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Category": {
      "type": "object",
      "properties": {
        "categoryID": {
          "format": "int32",
          "type": "integer"
        },
        "categoryName": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "product": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        }
      }
    },
...
    "CreateStyleModel": {
      "required": [
        "styleName"
      ],
      "type": "object",
      "properties": {
        "styleName": {
          "minLength": 1,
          "type": "string"
        }
      }
    },
    "CreateUserModel": {
      "required": [
        "userName"
      ],
      "type": "object",
      "properties": {
        "userName": {
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "roleId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "Feedback": {
      "type": "object",
      "properties": {
        "feedbackId": {
          "format": "int32",
          "type": "integer"
        },
        "rating": {
          "format": "int32",
          "type": "integer"
        },
        "userName": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "comment": {
          "type": "string"
        },
        "feedbackTime": {
          "format": "date-time",
          "type": "string"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "product": {
          "$ref": "#/definitions/Product"
        }
      }
    },
...
    "Order": {
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "totalPrice": {
          "format": "double",
          "type": "number"
        },
        "userName": {
          "type": "string"
        },
        "orderDate": {
          "format": "date-time",
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "orderDetail": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderDetail"
          }
        }
      }
    },
    "OrderDetail": {
      "type": "object",
      "properties": {
        "orderDetailId": {
          "format": "int32",
          "type": "integer"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "totalPrice": {
          "format": "double",
          "type": "number"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "order": {
          "$ref": "#/definitions/Order"
        },
        "product": {
          "$ref": "#/definitions/Product"
        }
      }
    },
    "Product": {
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "productCode": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "productName": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "stockQuantity": {
          "format": "int32",
          "type": "integer"
        },
        "categoryId": {
          "format": "int32",
          "type": "integer"
        },
        "styleId": {
          "format": "int32",
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "ingredient": {
          "type": "string"
        },
        "discount": {
          "format": "double",
          "type": "number"
        },
        "warning": {
          "type": "string"
        },
        "createTime": {
          "format": "date-time",
          "type": "string"
        },
        "volume": {
          "format": "double",
          "type": "number"
        },
        "feedbacks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Feedback"
          }
        },
        "category": {
          "$ref": "#/definitions/Category"
        },
        "style": {
          "$ref": "#/definitions/Style"
        },
        "orderDetail": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderDetail"
          }
        }
      }
    },
...
    "Role": {
      "type": "object",
      "properties": {
        "roleId": {
          "format": "int32",
          "type": "integer"
        },
        "roleName": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "user": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        }
      }
    },
    "Style": {
      "type": "object",
      "properties": {
        "styleId": {
          "format": "int32",
          "type": "integer"
        },
        "styleName": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        }
      }
    },
...
    "UpdateProductModel": {
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "productCode": {
          "type": "string"
        },
        "image": {
          "format": "binary",
          "type": "string"
        },
        "productName": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "stockQuantity": {
          "format": "int32",
          "type": "integer"
        },
        "discount": {
          "format": "double",
          "type": "number"
        },
        "ingredient": {
          "type": "string"
        },
        "warning": {
          "type": "string"
        },
        "categoryId": {
          "format": "int32",
          "type": "integer"
        },
        "styleId": {
          "format": "int32",
          "type": "integer"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "volume": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "UpdateStyleModel": {
      "required": [
        "styleName"
      ],
      "type": "object",
      "properties": {
        "styleName": {
          "minLength": 1,
          "type": "string"
        }
      }
    },
    "UpdateUserModel": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "roleId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "dob": {
          "format": "date-time",
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "roleId": {
          "format": "int32",
          "type": "integer"
        },
        "phoneNumber": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "createDate": {
          "format": "date-time",
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/Role"
        },
        "feedbacks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Feedback"
          }
        },
        "order": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Order"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}